+++ R/theme-elements.R
@@ -23,6 +23,7 @@
 #' @seealso [geom_nfl_logos()], [geom_nfl_headshots()], [geom_nfl_wordmarks()],
 #'   and [geom_from_path()] for more information on valid team abbreviations,
 #'   player IDs, and other parameters.
+#' @seealso The examples on <https://nflplotr.nflverse.com/articles/nflplotR.html>
 #' @return An S3 object of class `element`.
 #' @examples
 #' \donttest{
@@ -80,16 +81,51 @@
 #' ggplot(dfh, aes(x = player_gsis, y = random_value)) +
 #'   geom_col(width = 0.5) +
 #'   theme_minimal() +
-#'   theme(axis.text.x = element_nfl_headshot())
+#'   theme(axis.text.x = element_nfl_headshot(size = 1))
 #'
 #' # use headshots for y-axis
 #' ggplot(dfh, aes(y = player_gsis, x = random_value)) +
 #'   geom_col(width = 0.5) +
 #'   theme_minimal() +
-#'   theme(axis.text.y = element_nfl_headshot())
+#'   theme(axis.text.y = element_nfl_headshot(size = 1))
 #'
 #' # Restore old options
 #' options(old)
+#'
+#' #############################################################################
+#' # Wordmarks and other Images
+#' #############################################################################
+#'
+#' library(ggplot2)
+#'
+#' df <- mtcars |>
+#'   dplyr::mutate(
+#'     team = sample(c("LAC", "BUF", "DAL", "ARI"), nrow(mtcars), TRUE),
+#'     player = sample(
+#'     c("00-0033873", "00-0035228", "00-0036355", "00-0019596"),
+#'     nrow(mtcars),
+#'     TRUE
+#'     )
+#'   )
+#'
+#' ggplot(df, aes(x = mpg, y = disp)) +
+#'   geom_point() +
+#'   facet_wrap(vars(team)) +
+#'   labs(
+#'     title = tools::toTitleCase("These are random teams and data"),
+#'     subtitle = "I just want to show how the nflplotR theme elements work",
+#'     caption = "https://github.com/nflverse/nflseedR/raw/master/man/figures/caption.png"
+#'   ) +
+#'   theme_minimal() +
+#'   theme(
+#'     plot.title.position = "plot",
+#'     plot.title = element_text(face = "bold"),
+#'     axis.title = element_blank(),
+#'     # make wordmarks of team abbreviations
+#'     strip.text = element_nfl_wordmark(size = 1),
+#'     # load image from url in caption
+#'     plot.caption = element_path(hjust = 1, size = 0.4)
+#'   )
 #' }
 #' @name element
 #' @aliases NULL
@@ -142,14 +178,14 @@ element_path <- function(alpha = NULL, colour = NA, hjust = NULL, vjust = NULL,
 #' @export
 element_grob.element_nfl_logo <- function(element, label = "", x = NULL, y = NULL,
                                           alpha = NULL, colour = NULL,
-                                          hjust = NULL, vjust = NULL,
+                                          hjust = 0.5, vjust = 0.5,
                                           size = NULL, ...) {
 
   if (is.null(label)) return(ggplot2::zeroGrob())
 
   n <- max(length(x), length(y), 1)
-  vj <- vjust %||% element$vjust
-  hj <- hjust %||% element$hjust
+  vj <- element$vjust %||% vjust
+  hj <- element$hjust %||% hjust
   x <- x %||% unit(rep(hj, n), "npc")
   y <- y %||% unit(rep(vj, n), "npc")
   alpha <- alpha %||% element$alpha
@@ -181,14 +217,14 @@ element_grob.element_nfl_logo <- function(element, label = "", x = NULL, y = NUL
 #' @export
 element_grob.element_nfl_wordmark <- function(element, label = "", x = NULL, y = NULL,
                                               alpha = NULL, colour = NULL,
-                                              hjust = NULL, vjust = NULL,
+                                              hjust = 0.5, vjust = 0.5,
                                               size = NULL, ...) {
 
   if (is.null(label)) return(ggplot2::zeroGrob())
 
   n <- max(length(x), length(y), 1)
-  vj <- vjust %||% element$vjust
-  hj <- hjust %||% element$hjust
+  vj <- element$vjust %||% vjust
+  hj <- element$hjust %||% hjust
   x <- x %||% unit(rep(hj, n), "npc")
   y <- y %||% unit(rep(vj, n), "npc")
   alpha <- alpha %||% element$alpha
@@ -221,14 +257,14 @@ element_grob.element_nfl_wordmark <- function(element, label = "", x = NULL, y =
 #' @export
 element_grob.element_nfl_headshot <- function(element, label = "", x = NULL, y = NULL,
                                               alpha = NULL, colour = NULL,
-                                              hjust = NULL, vjust = NULL,
+                                              hjust = 0.5, vjust = 0.5,
                                               size = NULL, ...) {
 
   if (is.null(label)) return(ggplot2::zeroGrob())
 
   n <- max(length(x), length(y), 1)
-  vj <- vjust %||% element$vjust
-  hj <- hjust %||% element$hjust
+  vj <- element$vjust %||% vjust
+  hj <- element$hjust %||% hjust
   x <- x %||% unit(rep(hj, n), "npc")
   y <- y %||% unit(rep(vj, n), "npc")
   alpha <- alpha %||% element$alpha
@@ -261,16 +297,16 @@ element_grob.element_nfl_headshot <- function(element, label = "", x = NULL, y =
 #' @export
 element_grob.element_path <- function(element, label = "", x = NULL, y = NULL,
                                       alpha = NULL, colour = NULL,
-                                      hjust = NULL, vjust = NULL,
+                                      hjust = 0.5, vjust = 0.5,
                                       size = NULL, ...) {
 
   if (is.null(label)) return(ggplot2::zeroGrob())
 
   n <- max(length(x), length(y), 1)
-  vj <- vjust %||% element$vjust
-  hj <- hjust %||% element$hjust
-  x <- x %||% unit(rep(hj, n), "npc")
-  y <- y %||% unit(rep(vj, n), "npc")
+  vj <- element$vjust %||% vjust
+  hj <- element$hjust %||% hjust
+  x <- x %||% rep(hj, n)
+  y <- y %||% rep(vj, n)
   alpha <- alpha %||% element$alpha
   colour <- colour %||% rep(element$colour, n)
   size <- size %||% element$size
@@ -298,7 +334,7 @@ element_grob.element_path <- function(element, label = "", x = NULL, y = NULL,
   )
 }
 
-axisImageGrob <- function(i, label, alpha, colour, data, x, y, hjust, vjust,
+axisImageGrob <- function(i, label, alpha, colour, x, y, hjust, vjust,
                           width = 1, height = 1,
                           type = c("teams", "headshots", "wordmarks", "path")) {
   make_null <- FALSE
@@ -361,15 +397,18 @@ axisImageGrob <- function(i, label, alpha, colour, data, x, y, hjust, vjust,
     }
   }
 
-  grid::rasterGrob(
-    new,
-    x = x[i],
-    y = y[i],
-    width = grid::unit(width, "snpc"),
-    height = grid::unit(height, "snpc"),
-    hjust = hjust,
-    vjust = vjust
+  grid <- grid::rasterGrob(new, just = c(hjust, vjust))
+
+  grid$vp <- grid::viewport(
+    x = grid::unit(x[i], "npc"),
+    y = grid::unit(y[i], "npc"),
+    width = grid::unit(width, "npc"),
+    height = grid::unit(height, "npc")
+    # angle = data$angle[i],
   )
+
+  grid
+
 }
 
 #' @export